+++ src/aks-preview/azext_aks_preview/_validators.py
@@ -341,6 +341,30 @@ def validate_node_public_ip_tags(ns):
         ns.node_public_ip_tags = tags_dict
 
 
+def validate_egress_gtw_nodeselector(namespace):
+    """Validates that provided node selector is a valid format"""
+
+    if not hasattr(namespace, 'egx_gtw_nodeselector'):
+        return
+
+    labels = namespace.egx_gtw_nodeselector
+
+    if labels is None:
+        # no specify any labels
+        namespace.egx_gtw_nodeselector = {}
+        return
+
+    if isinstance(labels, list):
+        labels_dict = {}
+        for item in labels:
+            labels_dict.update(validate_label(item))
+        after_validation_labels = labels_dict
+    else:
+        after_validation_labels = validate_label(labels)
+
+    namespace.egx_gtw_nodeselector = after_validation_labels
+
+
 def validate_nodepool_labels(namespace):
     """Validates that provided node labels is a valid format"""
 
@@ -534,13 +558,6 @@ def validate_assign_kubelet_identity(namespace):
         if not is_valid_resource_id(namespace.assign_kubelet_identity):
             raise CLIError(
                 "--assign-kubelet-identity is not a valid Azure resource ID.")
-
-
-def validate_prompt_input(namespace):
-    if namespace.prompt is None:
-        return
-    if not re.search(r'[a-zA-Z]', namespace.prompt):
-        raise InvalidArgumentValueError('--prompt does not contain any alphabet character')
 
 
 def validate_snapshot_name(namespace):