+++ src/containerapp/azext_containerapp/_constants.py
@@ -27,7 +27,7 @@ CONTAINER_APP_EXTENSION_TYPE = "microsoft.app.environment"
 MANAGED_CERTIFICATE_RT = "managedCertificates"
 PRIVATE_CERTIFICATE_RT = "certificates"
 
-DEV_SERVICE_LIST = ["kafka", "postgres", "redis", "mariadb"]
+DEV_SERVICE_LIST = ["kafka", "postgres", "redis", "mariadb", "qdrant"]
 
 DEV_KAFKA_IMAGE = 'kafka'
 DEV_KAFKA_SERVICE_TYPE = 'kafka'
@@ -45,6 +45,10 @@ DEV_MARIADB_IMAGE = 'mariadb'
 DEV_MARIADB_SERVICE_TYPE = 'mariadb'
 DEV_MARIADB_CONTAINER_NAME = 'mariadb'
 
+DEV_QDRANT_IMAGE = 'qdrant'
+DEV_QDRANT_SERVICE_TYPE = 'qdrant'
+DEV_QDRANT_CONTAINER_NAME = 'qdrant'
+
 PENDING_STATUS = "Pending"
 SUCCEEDED_STATUS = "Succeeded"
 UPDATING_STATUS = "Updating"
@@ -61,6 +65,7 @@ APPLE_SECRET_SETTING_NAME = "apple-provider-authentication-secret"
 UNAUTHENTICATED_CLIENT_ACTION = ['RedirectToLoginPage', 'AllowAnonymous', 'Return401', 'Return403']
 FORWARD_PROXY_CONVENTION = ['NoProxy', 'Standard', 'Custom']
 CHECK_CERTIFICATE_NAME_AVAILABILITY_TYPE = "Microsoft.App/managedEnvironments/certificates"
+CONNECTED_ENV_CHECK_CERTIFICATE_NAME_AVAILABILITY_TYPE = "Microsoft.App/connectedEnvironments/certificates"
 
 NAME_INVALID = "Invalid"
 NAME_ALREADY_EXISTS = "AlreadyExists"