+++ src/containerapp/azext_containerapp/_help.py
@@ -445,6 +445,11 @@ helps['containerapp service mariadb'] = """
     short-summary: Commands to manage the mariadb service for the Container Apps environment.
 """
 
+helps['containerapp service qdrant'] = """
+    type: group
+    short-summary: Commands to manage the qdrant service for the Container Apps environment.
+"""
+
 helps['containerapp service redis create'] = """
     type: command
     short-summary: Command to create the redis service.
@@ -465,6 +470,11 @@ helps['containerapp service mariadb create'] = """
     short-summary: Command to create the mariadb service.
 """
 
+helps['containerapp service qdrant create'] = """
+    type: command
+    short-summary: Command to create the qdrant service.
+"""
+
 helps['containerapp service redis delete'] = """
     type: command
     short-summary: Command to delete the redis service.
@@ -483,6 +493,11 @@ helps['containerapp service kafka delete'] = """
 helps['containerapp service mariadb delete'] = """
     type: command
     short-summary: Command to delete the mariadb service.
+"""
+
+helps['containerapp service qdrant delete'] = """
+    type: command
+    short-summary: Command to delete the qdrant service.
 """
 
 helps['containerapp env update'] = """
@@ -1793,6 +1808,49 @@ helps['containerapp create'] = """
           az containerapp create -n MyContainerapp -g MyResourceGroup \\
               --image my-app:v1.0 --environment MyContainerappConnectedEnv \\
               --environment-type connected
+    - name: Create a container app from a new GitHub Actions workflow in the provided GitHub repository
+      text: |
+          az containerapp create -n MyContainerapp -g MyResourceGroup \\
+          --environment MyContainerappEnv --registry-server MyRegistryServer \\
+          --registry-user MyRegistryUser --registry-pass MyRegistryPass \\
+          --repo https://github.com/myAccount/myRepo
+    - name: Create a Container App from the provided application source
+      text: |
+          az containerapp create -n MyContainerapp -g MyResourceGroup \\
+          --environment MyContainerappEnv --registry-server MyRegistryServer \\
+          --registry-user MyRegistryUser --registry-pass MyRegistryPass \\
+          --source .
+"""
+
+# containerapp update for preview
+helps['containerapp update'] = """
+    type: command
+    short-summary: Update a container app. In multiple revisions mode, create a new revision based on the latest revision.
+    examples:
+    - name: Update a container app's container image.
+      text: |
+          az containerapp update -n MyContainerapp -g MyResourceGroup \\
+              --image myregistry.azurecr.io/my-app:v2.0
+    - name: Update a container app's resource requirements and scale limits.
+      text: |
+          az containerapp update -n MyContainerapp -g MyResourceGroup \\
+              --cpu 0.5 --memory 1.0Gi \\
+              --min-replicas 4 --max-replicas 8
+    - name: Update a container app with an http scale rule
+      text: |
+          az containerapp update -n myapp -g mygroup \\
+              --scale-rule-name my-http-rule \\
+              --scale-rule-http-concurrency 50
+    - name: Update a container app with a custom scale rule
+      text: |
+          az containerapp update -n myapp -g mygroup \\
+              --scale-rule-name my-custom-rule \\
+              --scale-rule-type my-custom-type \\
+              --scale-rule-metadata key=value key2=value2 \\
+              --scale-rule-auth triggerparam=secretref triggerparam=secretref
+    - name: Update a Container App from the provided application source
+      text: |
+          az containerapp update -n MyContainerapp -g MyResourceGroup --source .
 """
 
 # containerapp list for preview
@@ -1855,6 +1913,139 @@ helps['containerapp connected-env list'] = """
     - name: List connected environments by resource group.
       text: |
           az containerapp connected-env list -g MyResourceGroup
+"""
+
+helps['containerapp connected-env dapr-component'] = """
+    type: group
+    short-summary: Commands to manage Dapr components for Container Apps connected environments.
+"""
+
+helps['containerapp connected-env dapr-component list'] = """
+    type: command
+    short-summary: List Dapr components for a connected environment.
+    examples:
+    - name: List Dapr components for a connected environment.
+      text: |
+          az containerapp connected-env dapr-component list -g MyResourceGroup --name MyConnectedEnv
+"""
+
+helps['containerapp connected-env dapr-component show'] = """
+    type: command
+    short-summary: Show the details of a Dapr component.
+    examples:
+    - name: Show the details of a Dapr component.
+      text: |
+          az containerapp connected-env dapr-component show -g MyResourceGroup --dapr-component-name MyDaprComponentName --name MyConnectedEnv
+"""
+
+helps['containerapp connected-env dapr-component set'] = """
+    type: command
+    short-summary: Create or update a Dapr component.
+    examples:
+    - name: Create a Dapr component.
+      text: |
+          az containerapp connected-env dapr-component set -g MyResourceGroup --name MyEnv --yaml MyYAMLPath --dapr-component-name MyDaprComponentName
+"""
+
+helps['containerapp connected-env dapr-component remove'] = """
+    type: command
+    short-summary: Remove a Dapr component from a connected environment.
+    examples:
+    - name: Remove a Dapr component from a Container Apps connected environment.
+      text: |
+          az containerapp connected-env dapr-component remove -g MyResourceGroup --dapr-component-name MyDaprComponentName --name MyConnectedEnv
+"""
+
+helps['containerapp connected-env storage'] = """
+    type: group
+    short-summary: Commands to manage storage for the Container Apps connected environment.
+"""
+
+helps['containerapp connected-env storage list'] = """
+    type: command
+    short-summary: List the storages for a connected environment.
+    examples:
+    - name: List the storages for a connected environment.
+      text: |
+          az containerapp connected-env storage list -g MyResourceGroup -n MyConnectedEnv
+"""
+
+helps['containerapp connected-env storage show'] = """
+    type: command
+    short-summary: Show the details of a storage.
+    examples:
+    - name: Show the details of a storage.
+      text: |
+          az containerapp connected-env storage show -g MyResourceGroup --storage-name MyStorageName -n MyConnectedEnv
+"""
+
+helps['containerapp connected-env storage set'] = """
+    type: command
+    short-summary: Create or update a storage.
+    examples:
+    - name: Create a storage.
+      text: |
+          az containerapp connected-env storage set -g MyResourceGroup -n MyEnv --storage-name MyStorageName --access-mode ReadOnly --azure-file-account-key MyAccountKey --azure-file-account-name MyAccountName --azure-file-share-name MyShareName
+"""
+
+helps['containerapp connected-env storage remove'] = """
+    type: command
+    short-summary: Remove a storage from a connected environment.
+    examples:
+    - name: Remove a storage from a Container Apps connected environment.
+      text: |
+          az containerapp connected-env storage remove -g MyResourceGroup --storage-name MyStorageName -n MyConnectedEnv
+"""
+
+# Certificates Commands
+helps['containerapp connected-env certificate'] = """
+    type: group
+    short-summary: Commands to manage certificates for the Container Apps connected environment.
+"""
+
+helps['containerapp connected-env certificate list'] = """
+    type: command
+    short-summary: List certificates for a connected environment.
+    examples:
+    - name: List certificates for a connected environment.
+      text: |
+          az containerapp connected-env certificate list -g MyResourceGroup --name MyConnectedEnv
+    - name: List certificates by certificate id.
+      text: |
+          az containerapp connected-env certificate list -g MyResourceGroup --name MyConnectedEnv --certificate MyCertificateId
+    - name: List certificates by certificate name.
+      text: |
+          az containerapp connected-env certificate list -g MyResourceGroup --name MyConnectedEnv --certificate MyCertificateName
+    - name: List certificates by certificate thumbprint.
+      text: |
+          az containerapp connected-env certificate list -g MyResourceGroup --name MyConnectedEnv --thumbprint MyCertificateThumbprint
+"""
+
+helps['containerapp connected-env certificate upload'] = """
+    type: command
+    short-summary: Add or update a certificate.
+    examples:
+    - name: Add or update a certificate.
+      text: |
+          az containerapp connected-env certificate upload -g MyResourceGroup --name MyConnectedEnv --certificate-file MyFilepath
+    - name: Add or update a certificate with a user-provided certificate name.
+      text: |
+          az containerapp connected-env certificate upload -g MyResourceGroup --name MyConnectedEnv --certificate-file MyFilepath --certificate-name MyCertificateName
+"""
+
+helps['containerapp connected-env certificate delete'] = """
+    type: command
+    short-summary: Delete a certificate from the Container Apps connected environment.
+    examples:
+    - name: Delete a certificate from the Container Apps connected environment by certificate name
+      text: |
+          az containerapp connected-env certificate delete -g MyResourceGroup --name MyConnectedEnv --certificate MyCertificateName
+    - name: Delete a certificate from the Container Apps connected environment by certificate id
+      text: |
+          az containerapp connected-env certificate delete -g MyResourceGroup --name MyConnectedEnv --certificate MyCertificateId
+    - name: Delete a certificate from the Container Apps connected environment by certificate thumbprint
+      text: |
+          az containerapp connected-env certificate delete -g MyResourceGroup --name MyConnectedEnv --thumbprint MyCertificateThumbprint
 """
 
 # Container Apps Job Commands