+++ src/spring/azext_spring/custom.py
@@ -1341,7 +1341,7 @@ def storage_list_persistent_storage(client, resource_group, service, name):
 
 
 def certificate_add(cmd, client, resource_group, service, name, only_public_cert=None,
-                    vault_uri=None, vault_certificate_name=None, public_certificate_file=None):
+                    vault_uri=None, vault_certificate_name=None, public_certificate_file=None, enable_auto_sync=None):
     if vault_uri is None and public_certificate_file is None:
         raise InvalidArgumentValueError("One of --vault-uri and --public-certificate-file should be provided")
     if vault_uri is not None and public_certificate_file is not None:
@@ -1354,10 +1354,11 @@ def certificate_add(cmd, client, resource_group, service, name, only_public_cert
         if only_public_cert is None:
             only_public_cert = False
         properties = models.KeyVaultCertificateProperties(
-            type="KeyVaultCertificate",
+            type='KeyVaultCertificate',
             vault_uri=vault_uri,
             key_vault_cert_name=vault_certificate_name,
-            exclude_private_key=only_public_cert
+            exclude_private_key=only_public_cert,
+            auto_sync='Enabled' if enable_auto_sync is True else 'Disabled'
         )
     else:
         if os.path.exists(public_certificate_file):
@@ -1375,16 +1376,17 @@ def certificate_add(cmd, client, resource_group, service, name, only_public_cert
         )
     certificate_resource = models.CertificateResource(properties=properties)
 
-    def callback(pipeline_response, deserialized, headers):
-        return models.CertificateResource.deserialize(json.loads(pipeline_response.http_response.text()))
+    return client.certificates.begin_create_or_update(resource_group, service, name, certificate_resource)
 
-    return client.certificates.begin_create_or_update(
-        resource_group_name=resource_group,
-        service_name=service,
-        certificate_name=name,
-        certificate_resource=certificate_resource,
-        cls=callback
-    )
+
+def certificate_update(cmd, client, resource_group, service, name, enable_auto_sync=None):
+    certificate_resource = client.certificates.get(resource_group, service, name)
+
+    if certificate_resource.properties.type == 'KeyVaultCertificate':
+        if enable_auto_sync is not None:
+            certificate_resource.properties.auto_sync = 'Enabled' if enable_auto_sync is True else 'Disabled'
+
+    return client.certificates.begin_create_or_update(resource_group, service, name, certificate_resource)
 
 
 def certificate_show(cmd, client, resource_group, service, name):