+++ src/aks-preview/HISTORY.rst
@@ -12,10 +12,22 @@ To release a new version, please select a new version number (usually plus 1 to 
 Pending
 +++++++
 
-0.5.157
+0.5.160
 +++++++
 * Add `az aks machine list` command to fetch list of machines in an agentpool.
 * Add `az aks machine show` command to fetch information about a specific machine in an agentpool.
+
+0.5.159
++++++++
+* Revert `az aks copilot` Command
+
+0.5.158
++++++++
+* Add `enable-egress-gateway` and `disable-egress-gateway` to the `az aks mesh` command.
+
+0.5.157
++++++++
+* Add `--disable-workload-identity` to the `az aks update` command.
 
 0.5.156
 +++++++